requirements.txt
requests>=2.31.0
pandas>=2.2.2
openpyxl>=3.1.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.3
urllib3>=2.2.1